add support for flop computation
This change is backward compatible.  It adds an optional additional token to the skeleton description language, named "task_mode", which can be set to teh literal strings "time" or "flops".  If not specified, it defaults to "time", representing the previous behavior, to sleep for that amount of time.  If set to "flops", then the computation step will intepret the value of "task_length" as the number of flops to consume.

No attempt has been made to tune the skelelton compute emulation to represen an exact amount of flops.